rename NewTokenStandard to FungibleTokenStandard
