+++ src/FungibleTokenStandard.ts
