+++ src/side-loaded/run.ts
@@ -9,7 +9,7 @@ import {
   UInt64,
   UInt8,
 } from 'o1js';
-import { FungibleToken, VKeyMerkleMap } from '../NewTokenStandard.js';
+import { FungibleToken, VKeyMerkleMap } from '../FungibleTokenStandard.js';
 import {
   MintConfig,
   MintParams,
