+++ src/test/approve.test.ts
@@ -17,7 +17,7 @@ import {
   FungibleToken,
   FungibleTokenErrors,
   VKeyMerkleMap,
-} from '../NewTokenStandard.js';
+} from '../FungibleTokenStandard.js';
 import {
   MintConfig,
   MintParams,
