+++ src/test/transfer.test.ts
@@ -9,7 +9,7 @@ import {
   UInt8,
   VerificationKey,
 } from 'o1js';
-import { FungibleToken, VKeyMerkleMap } from '../NewTokenStandard.js';
+import { FungibleToken, VKeyMerkleMap } from '../FungibleTokenStandard.js';
 import {
   MintConfig,
   MintParams,
